merging latest (#5)
* Add support for dynamic commands

This patch adds support for commands which the parser may not know about
until runtime. This could be useful if we'd like to discover subcommands
at runtime, e.g. we're using a plugin system and have to enumerate our
plugin repository to know what commands are available.

* Revert "Add support for dynamic commands"

This was meant to go to a pull request, not straight on to master.

This reverts commit 997ea521c4394a51d88e67a9212ac2da5d96fd4f.

Co-authored-by: Casey Dahlin <sadmac@google.com>